+++ core/forms.py
@@ -1,6 +1,3 @@
-import functools
-import logging
-
 from django import forms
 from django.conf import settings
 from django.forms import ModelForm, HiddenInput, IntegerField, Form
@@ -10,9 +7,9 @@ from core.helper import form_utils, model_utils
 from core.helper import widgets_utils
 from core.models import CofkUnionComment, CofkUnionResource
 from manifestation.models import CofkUnionManifestation
-from person.models import CofkUnionPerson
+from person.person_utils import get_checked_form_url_by_pk
 from uploader.models import CofkUnionImage
-from work.models import CofkUnionWork
+from work.work_utils import get_checked_form_url_by_pk
 
 
 def build_search_components(sort_by_choices: list[tuple[str, str]]):
@@ -48,35 +45,10 @@ class RecrefForm(forms.Form):
         return ''  # tobe define by subclass
 
 
-def log_no_url(fn):
-    @functools.wraps(fn)
-    def _wrap(*args, **kwargs):
-        url = fn(*args, **kwargs)
-        if url is None:
-            logging.warning(f'{fn.__name__} failed, person not found [{args}]')
-            return ''
-
-        return url
-
-    return _wrap
-
-
-@log_no_url
-def get_peron_full_form_url_by_pk(pk):
-    if person := CofkUnionPerson.objects.get(pk=pk):
-        return reverse('person:full_form', args=[person.iperson_id])
-
-
-@log_no_url
-def get_work_full_form_url_by_pk(pk):
-    if work := CofkUnionWork.objects.get(pk=pk):
-        return reverse('work:full_form', args=[work.iwork_id])
-
-
 class PersonRecrefForm(RecrefForm):
     @property
     def target_url(self) -> str:
-        return get_peron_full_form_url_by_pk(self.initial.get('target_id'))
+        return get_checked_form_url_by_pk(self.initial.get('target_id'))
 
 
 class LocRecrefForm(RecrefForm):
@@ -88,7 +60,7 @@ class LocRecrefForm(RecrefForm):
 class WorkRecrefForm(RecrefForm):
     @property
     def target_url(self) -> str:
-        return get_work_full_form_url_by_pk(self.initial.get('target_id'))
+        return get_checked_form_url_by_pk(self.initial.get('target_id'))
 
 
 class ManifRecrefForm(RecrefForm):
