+++ core/helper/form_utils.py
@@ -1,13 +1,16 @@
 import logging
 import warnings
-from typing import Iterable
+from typing import Iterable, Type
 
 from django import forms
-from django.db.models import TextChoices
-from django.forms import BoundField
+from django.db.models import TextChoices, Choices, Model
+from django.forms import BoundField, CharField
 from django.template.loader import render_to_string
 
-from core.helper import widgets_utils
+from core.helper import widgets_utils, data_utils, view_utils
+from core.helper.common_recref_adapter import RecrefFormAdapter
+from core.models import Recref
+from person import person_utils
 
 log = logging.getLogger(__name__)
 
@@ -231,3 +234,124 @@ class CharSelectField(forms.CharField):
                          widget=forms.Select(choices=choices),
                          initial=choices[0][0],  # this can avoid invalid changed_data
                          **kwargs)
+
+
+class RelationField(CharField):
+
+    def __init__(self, choices_class: Type[Choices], *args, **kwargs):
+        self.choices_class = choices_class
+        widget = forms.CheckboxSelectMultiple(
+            choices=self.choices_class.choices
+        )
+        super().__init__(*args, required=False, widget=widget, **kwargs)
+
+    def prepare_value(self, value):
+        """
+        value: should be list of relation_types
+        """
+        value: list[str] = super().prepare_value(value)
+        selected_values = set(value)
+        possible_values = set(self.choices_class.values)
+
+        unexpected_values = selected_values - possible_values
+        if unexpected_values:
+            logging.warning(f'unexpected relationship_type [{unexpected_values}] ')
+
+        selected_values = selected_values & possible_values
+        return list(selected_values)
+
+    def clean(self, user_input_values):
+        return user_input_values  # return as list
+
+
+class UndefinedRelationChoices(TextChoices):
+    UNDEFINED = 'undefined', 'Undefined'
+
+
+class MultiRelRecrefForm(forms.Form):
+    template_name = 'core/component/multi_rel_recref_form.html'
+
+    name = forms.CharField(required=False)
+    target_id = forms.CharField(required=False, widget=forms.HiddenInput())
+    relationship_types = RelationField(UndefinedRelationChoices)
+
+    @classmethod
+    def create_recref_adapter(cls, *args, **kwargs) -> RecrefFormAdapter:
+        raise NotImplementedError()
+
+    @classmethod
+    def get_rel_type_choices_values(cls):
+        return cls.base_fields['relationship_types'].choices_class.values
+
+    @property
+    def target_url(self):
+        raise NotImplementedError()
+
+    @classmethod
+    def get_target_id(cls, recref: Recref):
+        raise NotImplementedError()
+
+    def find_recref_list_by_target_id(self, host_model: Model, target_id):
+        raise NotImplementedError()
+
+    def create_or_delete(self, host_model: Model, username):
+        """
+        create or delete
+        """
+        if not self.is_valid():
+            logging.warning(f'[{self.__class__.__name__}] do nothing is invalid')
+            return
+
+        recref_adapter = self.create_recref_adapter(host_model)
+        data: dict = self.cleaned_data
+        selected_rel_types = set(data['relationship_types'])
+
+        org_maps: list[Recref] = list(self.find_recref_list_by_target_id(host_model, data['target_id']))
+
+        # delete unchecked relation
+        _maps = (m for m in org_maps if m.relationship_type not in selected_rel_types)
+        for m in _maps:
+            log.info(f'delete [{m.relationship_type}][{m}]')
+            m.delete()
+
+        # add checked relation
+        new_types = selected_rel_types - {m.relationship_type for m in org_maps}
+        target_model = recref_adapter.find_target_instance(data['target_id'])
+        for new_type in new_types:
+            recref = recref_adapter.upsert_recref(new_type, host_model, target_model)
+            recref.update_current_user_timestamp(username)
+            recref.save()
+            log.info(f'add new [{target_model}][{recref}]')
+
+    @classmethod
+    def create_formset_by_records(cls, post_data,
+                                  records: Iterable[Recref], prefix):
+        initial_list = []
+        recref_adapter = cls.create_recref_adapter()
+        records = (r for r in records if r.relationship_type in cls.get_rel_type_choices_values())
+        for person_id, recref_list in data_utils.group_by(records, lambda r: cls.get_target_id(r)).items():
+            recref_list: list[Recref]
+            initial_list.append({
+                'name': recref_adapter.find_target_display_name_by_id(
+                    recref_adapter.get_target_id(recref_list[0])
+                ),
+                'target_id': person_id,
+                'relationship_types': {m.relationship_type for m in recref_list},
+            })
+
+        formset = view_utils.create_formset(
+            cls, post_data=post_data, prefix=prefix,
+            initial_list=initial_list,
+            extra=0,
+        )
+        return formset
+
+
+class TargetPersonMRRForm(MultiRelRecrefForm):
+    @property
+    def target_url(self):
+        return person_utils.get_checked_form_url_by_pk(self.initial.get('target_id'))
+
+    @classmethod
+    def get_target_id(cls, recref):
+        return recref.person_id
