+++ work/work_utils.py
@@ -5,6 +5,7 @@ from django.urls import reverse
 from core.constant import REL_TYPE_CREATED, REL_TYPE_WAS_ADDRESSED_TO, REL_TYPE_WAS_SENT_FROM, REL_TYPE_WAS_SENT_TO
 from location import location_utils
 from person import person_utils
+from siteedit2.utils.log_utils import log_no_url
 from work.models import CofkUnionWork
 
 
@@ -68,3 +69,9 @@ def get_form_url(iwork_id):
 
 def create_work_id(iwork_id) -> str:
     return f'cofk_union_work-iwork_id:{iwork_id}'
+
+
+@log_no_url
+def get_checked_form_url_by_pk(pk):
+    if work := CofkUnionWork.objects.get(pk=pk):
+        return reverse('work:full_form', args=[work.iwork_id])
